+++ src/services/AssetContext.js
@@ -1,3 +1,4 @@
 import { createContext } from 'react';
 
 export const AssetContext = createContext(null);
+export const UpdateAssetContext = createContext(null);
