--- src/types/HasUpdateAsset.ts
@@ -1,6 +1,0 @@
-import { Updater } from "use-immer";
-import Asset from "./Asset";
-
-export default interface hasUpdateAsset {
-	updateAsset: Updater<Asset>;
-}
